Add files via upload
Notice that one device is connected to pin D61 (A7) which is not a valid PCI-serviced pin.  This sketch still detects that device but simply can't read via Pin Change Interrupt.  This means you'll have to resort to other, more rudimentary means of reading a device on those type of pins.